[REFACTOR] Host -> Profile
