+++ src/index.js
@@ -4,7 +4,7 @@ const { port } = require('../config')
 const db = require('./db')
 const placesRoutes = require('./routes/places')
 const authRoutes = require('./routes/auth')
-const hostsRoutes = require('./routes/hosts')
+const profileRoutes = require('./routes/profile')
 
 const app = express()
 
@@ -22,7 +22,7 @@ app.get('/', (req, res, next) => {
 // ? Routes
 placesRoutes(app)
 authRoutes(app)
-hostsRoutes(app)
+profileRoutes(app)
 
 app.listen(port, () => {
   console.log(`Listening on: http://localhost:${port}`)
