--- src/services/hosts.js
@@ -1,19 +1,0 @@
-const HostsSchema = require('../utils/schemas/hosts')
-
-const createHost = ({ host }) => {
-  const myHost = new HostsSchema(host)
-
-  return new Promise((resolve, reject) => {
-    myHost.save((error, createdHost) => {
-      if (error) {
-        reject(error)
-      }
-
-      resolve(createdHost.id)
-    })
-  })
-}
-
-module.exports = {
-  createHost
-}
